requirements.txt
pillow==2.2.1